Moves Haskell parser commify and strip_comments to its parser.
