+++ Cargo.toml
@@ -10,6 +10,8 @@ authors = [
 regex = "0.2.1"
 maplit = "*"
 walkdir = "1"
-corroder-parser = { path = "./parser" }
 base64 = "0.5.2"
 lalrpop-util = "0.12.5"
+
+[dependencies.parser-haskell]
+path = "./parser-haskell"
