+++ parser-haskell/.gitignore
@@ -1,1 +1,1 @@
-src/calculator.rs
+src/haskell.rs
