+++ parser-haskell/build.rs
@@ -1,6 +1,6 @@
 extern crate lalrpop;
 
 fn main() {
-    println!("cargo:rerun-if-changed=src/calculator.lalrpop");
+    println!("cargo:rerun-if-changed=src/haskell.lalrpop");
     lalrpop::process_root().unwrap();
 }
