+++ src/main.rs
@@ -1,16 +1,15 @@
 #[macro_use] extern crate maplit;
 extern crate base64;
-extern crate corroder_parser;
+extern crate parser_haskell;
 extern crate lalrpop_util;
 extern crate regex;
 extern crate walkdir;
 
-use corroder_parser::ast;
-use corroder_parser::ast::{Expr, Pat, Ty};
-use corroder_parser::calculator;
-use corroder_parser::util::{print_parse_error, simplify_parse_error};
+use parser_haskell::ast;
+use parser_haskell::ast::{Expr, Pat, Ty};
+use parser_haskell::util::{print_parse_error, simplify_parse_error};
 
-use regex::{Regex, Captures};
+use regex::Regex;
 use std::borrow::Borrow;
 use std::io::prelude::*;
 use std::fs::{File};
@@ -18,228 +17,6 @@ use std::env;
 use std::collections::BTreeSet;
 
 use walkdir::WalkDir;
-
-// TODO rename "strip comments and other stuff too i guess"
-fn strip_comments(text: &str) -> String {
-    // Strip comments
-    let re = Regex::new(r"--[^\n\r]*").unwrap();
-    let text = re.replace_all(&text, "").to_string();
-    let re = Regex::new(r"\{-[\s\S]*?-\}").unwrap();
-    let text = re.replace_all(&text, "").to_string();
-
-    // Strip trailing semicolons (so we don't have "empty statements")
-    let re = Regex::new(r"(?m);+\s*$").unwrap();
-    let text = re.replace_all(&text, "").to_string();
-
-    // Strip preprocessor decls
-    let re = Regex::new(r"(?m)^#(if|ifn?def|endif|else).*").unwrap();
-    let text = re.replace_all(&text, "").to_string();
-
-    // TODO this should be handled in the parser
-    let escape_re = Regex::new(r#"\\([nrt'"\\0]|NUL|ESC)"#).unwrap();
-    let decode_escapes = |text: &str| {
-        let text = escape_re.replace_all(text, |caps: &Captures| {
-            match &caps[1] {
-                "n" => "\n",
-                "r" => "\r",
-                "t" => "\t",
-                "'" => "'",
-                "\"" => "\"",
-                "\\" => "\\",
-                "0" | "NUL" => "\0",
-                "ESC" => "\x1b",
-                s => panic!("str escape {}", s),
-            }.into()
-        });
-        text.to_string()
-    };
-
-    // Replace all strings with a base64 encoded version to make the parser simpler.
-    // If its possible to get LALRPOP to not complain with proper string regexes, should just use
-    // that instead
-    let re = Regex::new(r#""(([^"\\]|\\.)*?)""#).unwrap();
-    let text = re.replace_all(&text, |caps: &Captures| {
-        let v = decode_escapes(&caps[1]);
-        format!("\"{}\"", base64::encode(&v))
-    }).to_string();
-
-    // Char literals.
-    let re = Regex::new(r"'([^'\\]|\\.{1,3})'").unwrap();
-    let text = re.replace_all(&text, |caps: &Captures| {
-        let v = decode_escapes(&caps[1]);
-        assert!(v.len() == 1, "multi char literal {:?}", v);
-        format!("'{}'", base64::encode(&v))
-    }).to_string();
-
-    text
-}
-
-fn decode_literal(s: &str) -> String {
-    let vec = base64::decode(s).expect("invalid base64");
-    String::from_utf8(vec).expect("invalid UTF-8")
-}
-
-fn word_is_block_word(word: &str) -> bool {
-    word == "do" || word == "where" || word == "of" || word == "let"
-}
-
-/// Convert indentation to semicolon-delimited brackets, so it can be parsed more easily.
-fn commify(val: &str) -> String {
-    let re_space = Regex::new(r#"^[ \t]+"#).unwrap();
-    let re_nl = Regex::new(r#"^\r?\n"#).unwrap();
-    let re_word = Regex::new(r#"([\(\{\[\]\}\)]|[^ \t\r\n\(\{\[\]\}\)]+)"#).unwrap();
-
-    let mut out = String::new();
-
-    // Previous indentation levels
-    let mut stash: Vec<usize> = vec![];
-    // Previous brace nesting levels.
-    let mut braces: Vec<isize> = vec![];
-    // Previous word was a block starting word, option containing its indent level.
-    let mut trigger = None;
-    // How many spaces to indent.
-    let mut indent = 0;
-    // Check if this is the first word in the line.
-    let mut first = true;
-
-    let mut commentless = strip_comments(val);
-
-    // HACK Until the indentation logic below is fixed (unsure what the right approach is), we need these
-    // C.hs
-    commentless = commentless.replace("let lhsvar", "let\n                    lhsvar");
-    commentless = commentless.replace("let allow_signed", "let\n            allow_signed");
-    commentless = commentless.replace("let isDefault", "let\n        isDefault");
-    commentless = commentless.replace("let returnValue", "let\n                    returnValue");
-    // DeclAnalysis.hs
-    commentless = commentless.replace("of ShortMod", "of\n                                     ShortMod");
-    // CFG.hs
-    commentless = commentless.replace("let (returns", "let\n        (returns");
-
-    let mut v: &str = &commentless;
-    while v.len() > 0 {
-        if let Some(cap) = re_space.captures(v) {
-            let word = &cap[0];
-            out.push_str(word);
-            v = &v[word.len()..];
-
-            indent += word.len();
-        } else if let Some(cap) = re_nl.captures(v) {
-            let word = &cap[0];
-            out.push_str(word);
-            v = &v[word.len()..];
-
-            indent = 0;
-            first = true;
-            if stash.len() > 1 {
-                for _ in &stash[1..] {
-                    out.push_str(" ");
-                }
-            }
-        } else if let Some(cap) = re_word.captures(v) {
-            let word = &cap[0];
-
-            if first {
-                while {
-                    if let Some(last_level) = stash.last().map(|x| *x) {
-                        // Check if we decreased our indent level
-                        last_level > indent
-                    } else {
-                        false
-                    }
-                } {
-                    // out.push_str(&format!("[{:?}{:?}]", last_level, stash.last()));
-                    stash.pop();
-                    braces.pop();
-                    out.push_str("}");
-                }
-
-                if let Some(i) = stash.last() {
-                    if *i == indent && trigger.is_none() {
-                        out.push_str(";");
-                    }
-                }
-            }
-
-            if ["]", ")", "}"].contains(&word) {
-                if let Some(brace) = braces.last_mut() {
-                    *brace -= 1;
-                }
-            }
-            if ["[", "(", "{"].contains(&word) {
-                if let Some(brace) = braces.last_mut() {
-                    *brace += 1;
-                }
-            }
-
-            // End braces insertion when meeting an unbalanced ending ), }, or ]
-            while {
-                if let Some(brace) = braces.last().map(|x| *x) {
-                    brace < 0
-                } else {
-                    false
-                }
-            } {
-                stash.pop();
-                braces.pop();
-                out.push_str("}");
-            }
-
-            out.push_str(word);
-            v = &v[word.len()..];
-
-            if trigger.is_some() {
-                // Determine which column to start determining whitespace. Where the first word is
-                // or where the keyword is?
-                if first {
-                    stash.push(indent);
-                } else {
-                    stash.push(trigger.unwrap());
-                }
-            }
-            first = false;
-
-            trigger = if word_is_block_word(word) { Some(indent) } else { None };
-            if trigger.is_some() {
-                out.push_str("{");
-
-                // Trace brace indentation level.
-                braces.push(0);
-            }
-
-            indent += word.len();
-        } else {
-            unreachable!("unknown prop {:?}", v);
-        }
-    }
-    for _ in 0..stash.len() {
-        out.push_str("}");
-    }
-
-
-    // Replace trailing commas after where statements
-    // TODO fix this in the parser instead
-    let re = Regex::new(r#"where\s+;"#).unwrap();
-    let out = re.replace_all(&out, r#"where "#).to_string();
-    // let re = Regex::new(r#"\};\s*where\b"#).unwrap();
-    // let out = re.replace_all(&out, r#"} where"#).to_string();
-    let re = Regex::new(r#"\};\}"#).unwrap();
-    let out = re.replace_all(&out, r#"}}"#).to_string();
-
-    out
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 #[derive(Clone, Copy)]
 struct PrintState {
@@ -661,13 +438,13 @@ fn test_single_file() {
     if a.ends_with(".lhs") {
         contents = fix_lhs(&contents);
     }
-    let input = commify(&contents);
+    let contents = parser_haskell::preprocess(contents);
 
     let mut a = ::std::fs::File::create("temp.txt").unwrap();
-    a.write_all(input.as_bytes());
+    a.write_all(contents.as_bytes());
 
     let mut errors = Vec::new();
-    match calculator::parse_Module(&mut errors, &input) {
+    match parser_haskell::parse(&mut errors, &input) {
         Ok(okay) => println!("{:#?}", okay),
         Err(e) => {
             let e = simplify_parse_error(e);
@@ -734,14 +511,14 @@ fn test_no_regressions() {
         if path.ends_with(".lhs") {
             contents = fix_lhs(&contents);
         }
-        let input = commify(&contents);
 
         // Do not output test.txt
         //let mut a = ::std::fs::File::create("temp.txt").unwrap();
         //a.write_all(input.as_bytes());
 
         let mut errors = Vec::new();
-        match calculator::parse_Module(&mut errors, &input) {
+        let contents = parser_haskell::preprocess(contents);
+        match parser_haskell::parse(&mut errors, &contents) {
             Ok(v) => {
             }
             Err(e) => {
@@ -795,10 +572,10 @@ fn main() {
         if do_fix_lhs {
             contents = fix_lhs(&contents);
         }
+        let contents = parser_haskell::preprocess(&contents);
 
-        let input = commify(&contents);
         let mut errors = Vec::new();
-        match calculator::parse_Module(&mut errors, &input) {
+        match parser_haskell::parse(&mut errors, &contents) {
             Ok(v) => {
                 // println!("{:?}", p);
                 // continue;
@@ -809,7 +586,7 @@ fn main() {
             }
             Err(e) => {
                 println!("/* ERROR: cannot yet convert file {:?}", p);
-                print_parse_error(&input, &simplify_parse_error(e));
+                print_parse_error(&contents, &simplify_parse_error(e));
                 println!("*/");
             }
         }
